pyproject.toml
[tool.poetry]
name = "spinayarn"
version = "0.1.0"
description = "Generate Intearctive Stories with LLMs"
authors = ["bk62"]
license = "MIT"
readme = "README.md"
repository = "https://www.github.com.com/bk62/spinayarn"

[tool.poetry.dependencies]
python = "^3.10"
# llama-cpp-python = "^0.1.62"
pyyaml = "^6.0"
pydantic = "^1.10.9"
langchain = "^0.0.202"
jinja2 = "^3.1.2"
networkx = "^3.1"
tqdm = "^4.65.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.272"
black = "^23.3.0"
mypy = "^1.3.0"
types-pyyaml = "^6.0.12.10"
matplotlib = "^3.7.1"
jupyter = "^1.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
